fix: rename and fix icons problem
