+++ packages/pillar-icons/src/icons/action/index.ts
@@ -22,7 +22,7 @@ export { default as CircleDot } from './circleDot'
 export { default as CircleDotted } from './circleDotted'
 export { default as CircleInfo } from './circleInfo'
 export { default as CircleMinus } from './circleMinus'
-export { default as CircleMutliply } from './circleX'
+export { default as CircleX } from './circleX'
 export { default as CircleRefresh } from './circleRefresh'
 export { default as CircleWarning } from './circleWarning'
 export { default as Close } from './close'
@@ -64,7 +64,7 @@ export { default as SquareCheck } from './squareCheck'
 export { default as SquareDot } from './squareDot'
 export { default as SquareInfo } from './squareInfo'
 export { default as SquareMinus } from './squareMinus'
-export { default as SquareMultiply } from './squareX'
+export { default as SquareX } from './squareX'
 export { default as Star } from './star'
 export { default as StarOff } from './starOff'
 export { default as Stars } from './stars'
