+++ packages/pillar-icons/src/icons/action/users.tsx
@@ -1,10 +1,10 @@
 import { svgProps } from '../../shared'
 import { SvgType } from '../../type'
 
-const user = (props: SvgType) => (
+const users = (props: SvgType) => (
   <svg {...svgProps} {...props}>
-    <path d="M6 21v-2a4 4 0 0 1 4-4h4a4 4 0 0 1 4 4v2M16 7a4 4 0 1 1-8 0 4 4 0 0 1 8 0Z" />
+    <path d="M3 21v-2a4 4 0 0 1 4-4h4a4 4 0 0 1 4 4v2m1-17.9a4 4 0 0 1 0 7.8M21 21v-2a4 4 0 0 0-3-3.8M13 7a4 4 0 1 1-8 0 4 4 0 0 1 8 0Z" />
   </svg>
 )
 
-export default user
+export default users
